Supported making 64-bit and 32-bit builds with different NDK versions

+++ source/build-tdlib.sh
@@ -4,6 +4,7 @@ TDLIB_SOURCE_DIR=${1:-td}
 TDLIB_INSTALL_DIR=${2:-build/td}
 OPENSSL_INSTALL_DIR=${3:-build/openssl}
 ANDROID_STL=${4:-c++_shared}
+TDLIB_BUILD_SCRIPT="$(pwd)/build-tdlib-impl.sh"
 
 source "$(pwd)/setup.sh" --light
 
@@ -37,16 +38,21 @@ if [ -e "$TDLIB_INSTALL_DIR" ] ; then
   exit 1
 fi
 
+if [ ! -e "$TDLIB_BUILD_SCRIPT" ] ; then
+  echo "Error: file or directory \"$TDLIB_BUILD_SCRIPT\" doesn't exists."
+  exit 1
+fi
+
 ANDROID_SDK_ROOT="$(cd "$(dirname -- "$ANDROID_SDK_ROOT")" >/dev/null; pwd -P)/$(basename -- "$ANDROID_SDK_ROOT")"
-ANDROID_NDK_ROOT="$ANDROID_SDK_ROOT/ndk/$ANDROID_NDK_VERSION"
 TDLIB_SOURCE_DIR="$(cd "$(dirname -- "$TDLIB_SOURCE_DIR")" >/dev/null; pwd -P)/$(basename -- "$TDLIB_SOURCE_DIR")"
 TDLIB_INSTALL_DIR="$(cd "$(dirname -- "$TDLIB_INSTALL_DIR")" >/dev/null; pwd -P)/$(basename -- "$TDLIB_INSTALL_DIR")"
 OPENSSL_INSTALL_DIR="$(cd "$(dirname -- "$OPENSSL_INSTALL_DIR")" >/dev/null; pwd -P)/$(basename -- "$OPENSSL_INSTALL_DIR")"
 
 pushd "$TDLIB_SOURCE_DIR/example/android" > /dev/null || exit 1
 rm -rf build-native build-arm64-v8a build-armeabi-v7a build-x86_64 build-x86 tdlib
-sed -i.bak 's/cmake -DCMAKE_TOOLCHAIN_FILE/cmake -DCMAKE_SHARED_LINKER_FLAGS="-Wl,--gc-sections,--icf=safe" -DCMAKE_TOOLCHAIN_FILE/g' build-tdlib.sh || exit 1
-./build-tdlib.sh "$ANDROID_SDK_ROOT" "$NDK_VERSION" "$OPENSSL_INSTALL_DIR" "$ANDROID_STL" || exit 1
+rm build-tdlib.sh
+cp "$TDLIB_BUILD_SCRIPT" build-tdlib.sh
+./build-tdlib.sh "$ANDROID_SDK_ROOT" "$ANDROID_NDK_VERSION_PRIMARY" "$ANDROID_NDK_VERSION_LEGACY" "$OPENSSL_INSTALL_DIR" "$ANDROID_STL" || exit 1
 popd > /dev/null
 
 pushd "$TDLIB_SOURCE_DIR" > /dev/null || exit 1
